Version 3
Nothing big, some bug fixes and minor changes.
Removed unused table locking functions which seemed to prevent edit to work.
Added a delay when calling init between every unit start so it looks better. Can be modified.
Advanced computers do not start multishell by default in agetty. This is because of --noclear argument not working with it.